+++ openfisca_core/projectors/helpers.py
@@ -1,10 +1,11 @@
 from __future__ import annotations
 
 from collections.abc import Mapping
+from openfisca_core.entities.typing import Entity, GroupEntity, Role
 
 from openfisca_core import entities, projectors
 
-from .typing import Entity, GroupEntity, GroupPopulation, Population
+from .typing import GroupPopulation, Population
 
 
 def projectable(function):
@@ -124,7 +125,7 @@ def get_projector_from_shortcut(
         return projectors.FirstPersonToEntityProjector(population, parent)
 
     if isinstance(entity, entities.GroupEntity):
-        role: entities.Role | None = entities.find_role(entity.roles, shortcut, total=1)
+        role: Role | None = entities.find_role(entity.roles, shortcut, total=1)
 
         if role is not None:
             return projectors.UniqueRoleToEntityProjector(population, role, parent)